combine the vschema commands
